requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.34.0
    # via -r requirements.in
anyio==4.4.0
    # via
    #   anthropic
    #   groq
    #   httpx
    #   starlette
cachetools==5.4.0
    # via google-auth
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   pinecone-client
    #   requests
cffi==1.17.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   nltk
    #   uvicorn
cryptography==43.0.0
    # via pyjwt
deprecated==1.2.14
    # via pygithub
distro==1.9.0
    # via
    #   anthropic
    #   groq
dnspython==2.6.1
    # via
    #   -r requirements.in
    #   pymongo
fastapi==0.112.1
    # via -r requirements.in
filelock==3.15.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
    #   triton
flask==2.0.1
    # via -r requirements.in
fsspec==2024.6.1
    # via
    #   huggingface-hub
    #   torch
google-api-core==2.19.1
    # via google-api-python-client
google-api-python-client==2.141.0
    # via -r requirements.in
google-auth==2.33.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
googleapis-common-protos==1.63.2
    # via google-api-core
groq==0.9.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httplib2==0.22.0
    # via
    #   -r requirements.in
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.27.0
    # via
    #   anthropic
    #   groq
huggingface-hub==0.24.5
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via
    #   flask
    #   torch
jiter==0.5.0
    # via anthropic
joblib==1.4.2
    # via
    #   -r requirements.in
    #   nltk
    #   scikit-learn
loguru==0.7.2
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.3
    # via torch
nltk==3.8.1
    # via -r requirements.in
numpy==2.0.1
    # via
    #   -r requirements.in
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.20
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.6.37
    # via torch
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pillow==10.4.0
    # via sentence-transformers
pinecone-client==5.0.1
    # via -r requirements.in
pinecone-plugin-inference==1.0.3
    # via pinecone-client
pinecone-plugin-interface==0.0.7
    # via
    #   pinecone-client
    #   pinecone-plugin-inference
proto-plus==1.24.0
    # via google-api-core
protobuf==5.27.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pydantic==2.8.2
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   groq
pydantic-core==2.20.1
    # via pydantic
pygithub==2.3.0
    # via -r requirements.in
pygments==2.18.0
    # via rich
pyjwt[crypto]==2.9.0
    # via pygithub
pymongo==4.8.0
    # via -r requirements.in
pynacl==1.5.0
    # via pygithub
pyparsing==3.1.2
    # via httplib2
python-dateutil==2.9.0.post0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   transformers
redis==5.0.8
    # via -r requirements.in
regex==2024.7.24
    # via
    #   nltk
    #   transformers
requests==2.32.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   huggingface-hub
    #   pygithub
    #   transformers
rich==13.7.1
    # via -r requirements.in
rsa==4.9
    # via google-auth
safetensors==0.4.4
    # via transformers
scikit-learn==1.5.1
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.14.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==3.0.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   httpx
starlette==0.38.2
    # via fastapi
sympy==1.13.2
    # via torch
threadpoolctl==3.5.0
    # via
    #   -r requirements.in
    #   scikit-learn
tokenizers==0.19.1
    # via
    #   anthropic
    #   transformers
torch==2.4.0
    # via sentence-transformers
tqdm==4.66.5
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   nltk
    #   pinecone-client
    #   sentence-transformers
    #   transformers
transformers==4.44.0
    # via sentence-transformers
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   groq
    #   huggingface-hub
    #   pinecone-client
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   torch
uritemplate==4.1.1
    # via
    #   -r requirements.in
    #   google-api-python-client
urllib3==2.2.2
    # via
    #   pinecone-client
    #   pygithub
    #   requests
uvicorn==0.30.6
    # via -r requirements.in
werkzeug==2.0.1
    # via
    #   -r requirements.in
    #   flask
wrapt==1.16.0
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
